api/pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = [
  "pre-commit>=4.5.1,<5",
  "ruff>=0.15.0,<0.16",
  "watchdog>=6.0.0,<7",
  "python-dotenv>=1.2.1,<2",
  "pytest>=9.0.2,<10",
  "pytest-mock>=3.15.1,<4",
  "coveralls>=4.0.2,<5",
  "alembic>=1.18.3,<2",
  "pytest-asyncio>=1.3.0,<1.4",
  "mkdocs-material>=9.7.1",
  "mkdocs-material[imaging]>=9.7.1",
  "vulture>=2.14",
  "safety>=3.7.0",
  "bandit>=1.9.3",
  "pyrefly>=0.49.0"
]

[project]
authors = [{name = "Anderson Resende", email = "anderson@vinta.com.br"}]
dependencies = [
  "fastapi[standard]>=0.128.1,<0.129",
  "asyncpg>=0.31.0,<0.32",
  "fastapi-users[sqlalchemy]>=15.0.4,<16",
  "pydantic-settings>=2.12.0,<3",
  "fastapi-mail>=1.6.1,<2",
  "fastapi-pagination==0.15.9"
]
description = ""
name = "app"
readme = "README.md"
requires-python = ">=3.12,<3.13"
version = "0.0.6"

[tool.hatch.build.targets.sdist]
include = ["commands"]

[tool.hatch.build.targets.wheel]
include = ["commands"]

[tool.pyrefly]
ignore-errors-in-generated-code = true
project-includes = ["."]
python-version = "3.12"
search-path = ["."]

[tool.pyrefly.errors]
bad-assignment = false
invalid-argument = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
ignore = [
  "E501",  # line too long, handled by black
  "B008",  # do not perform function calls in argument defaults (FastAPI Depends)
  "W191"  # indentation contains tabs
]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "ARG001"  # unused arguments in functions
]

[tool.ruff.lint.per-file-ignores]
"alembic_migrations/versions/*" = ["ARG001"]
"app/modules/users/manager.py" = ["ARG001"]  # request argument required by signature
# Ignore check for unused arguments in tests/migrations that need specific signatures
"tests/**/*" = ["ARG001"]

[tool.uv]
package = false
